Solve challenge 3
Wow, my English detection heuristic works really well apparently. It
decoded the sentence "Cooking MC's like a pound of bacon" on the first
try!

Also, this is not nearly as slow as I expected it to be because of the
immutable sorted map. Scala and sbt overhead still dwarfs execution
time.